% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICschedule.R
\name{ICSchedule}
\alias{ICSchedule}
\title{Generates interval-censored Failure Time data with schedule visits}
\usage{
ICSchedule(n, shape, scale, visits, p, X = as.matrix(0),
  beta = as.matrix(0))
}
\description{
Generates interval-censored Failure Time data with schedule visits
}
\examples{
ICsimple(50,0.8,0.8)
}
\keyword{censored}
\keyword{interval}
