% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricI.R
\name{ricI}
\alias{ricI}
\title{Generates Case II interval-censored Failure Time data}
\usage{
ricI(n, shape, scale, pc, X = as.matrix(0), beta = as.matrix(0))
}
\description{
Generates Case II interval-censored Failure Time data
}
\examples{
ICsimple(50,0.8,0.8)
}
\keyword{censored}
\keyword{interval}
